requirement.txt
discord.py==2.3.2
Flask==2.3.3
pytz==2023.3
python-dotenv==1.0.0